[dagit] Repair endless scrolling Gaant
Summary:
Resolves #3063

On a Run page, we are currently depending on filtered logs to determine the metadata for the Gantt chart, which means that if you filter out "Event", we no longer have our start and end times in the metadata, and the chart scrolls forever even after the run is complete.

Fix this by providing all of the nodes for metadata computation, rather than just the filtered nodes. I think this was probably an oversight when I made changes to log filtering a couple months ago.

Test Plan: Kick off a run, unselect "Event" in the logs toolbar. Verify that the Gaant shows the correct execution, and stops properly when the run completes.

Reviewers: sashank, alangenfeld, yuhan, catherinewu

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D5255