Refactor environment_configs.py to share config construction code amongst configurable definitions
Summary:
I am prototyping a cleanup of the configured/config mapping stack and whilst I'm doing
that I'm picking off refactors that will make that easier/more scoped and that are code quality wins.

In this case, I want to consolidate the construction of config fields associated with anything that implements `ConfigurableMixin`.

Test Plan: BK

Reviewers: sandyryza, sashank, cdecarolis

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D5272