[sensors-10] schedule def new API
Summary:
Should note:  this will require a wipe for folks who are running off of master.  This changes the schema by dropping run_key as a column, and changing the tick data schema.

Thought about doing a migration, but it might save some complexity to require the wipe.

Test Plan: bk

Reviewers: dgibson, schrockn, sashank

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5282