[dagit] Properly disable LaunchButton
Summary:
Resolves #3324

The "disabled" state of the Run launch button doesn't get passed down to the button itself, so even if the button looks disabled, it's not. Fix this.

Test Plan: Force `disabled` prop on the `LaunchButtonDropdown` component via React devtools, verify that the button is disabled and does nothing when clicked.

Reviewers: bengotow, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D5393