Be able to use the full spectrum of config schema in composites
Summary:
While working on issue 3332 I came across the fact
that we don't support the full spectrum of configuration in
composite solids. For example, you could not have config be a bare int.
We required a dictionary. This was an artifcial limitation and
made this inconsitent with other configurable definitions

Test Plan: BK

Reviewers: sandyryza, cdecarolis

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D5567