[dagit] Restore Runs page countdown with loop fix
Summary:
With the addition of `notifyOnNetworkStatusChange` in `useCursorPaginatedQuery`, we can end up in an unfortunate situation where queries that load to partial refetches of the paginated query can cause a loop, toggling back and forth between "loading" and "refetching". This showed up on the Runs list:

- Load the Runs list.
- Hover over a status tag. This kicks off a query for stats on that run.
- This data led to a refetch of the main root runs query, which then looped.
- The loop effectively crashed the page.

It's really hard to figure out what exactly happened here because it's difficult to introspect the Apollo behavior, but it appeared that when the stats query wrote to the Apollo cache, this led to the root query re-fetching to synchronize the run object with the new stats data. For some reason the refetch did not reconcile successfully, thus beginning a loop to endlessly reconcile the data. The root query asks for `startTime` end `endTime` on the `PipelineRunStatsSnapshot` object, while `RunStatsQuery` asks for a completely different set of fields, so perhaps they were overwriting each other.

When the stats query used `fetchPolicy: 'no-cache'`, this didn't occur -- the root query refetch was not triggered. This implied that the `RunStatsQuery` writing to the cache was the catalyst, and if we could either avoid writing to the cache (not great) or avoid triggering the root query refetch while still writing to the cache, the bug could be resolved.

My hunch here was that it might work to add an `id` field to the stats object, hopefully caching it independently of the root query and allowing the root query not to have to perform the partial refetch. This appears to have worked. The `PipelineRunStatsSnapshot` object is now cached separately with its ID, and updates to that object are written there without triggering the root refetch.

This diff adds the countdown back to the page, along with the ID changes.

Going forward, we probably need to be more aggressive about giving "dependent" objects like these their own IDs so that we can avoid these kinds of cache problems.

Test Plan: View Runs page, hover over a status tag. Verify that the refetch loop does not occur.

Reviewers: bengotow, dgibson, prha, johann

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5509