[dagit] Pause Gaant when websocket disconnects
Summary:
Resolves #3015

Try to resolve confusion around in-progress runs when the WebSocket connection is lost, by pausing the Gaant chart and displaying a message when the WebSocket is not open.

Test Plan: Kick off a run, then kill the Dagit server. Verify that the warning appears and the gantt chart is paused.

Reviewers: bengotow, prha, sashank

Reviewed By: bengotow, prha

Differential Revision: https://dagster.phacility.com/D5620