[dagit] Clear the countdown interval on effect return
Summary: The countdown interval is hanging around after navigation from a page that uses the countdown, leading to React warnings about updating state on unmounted components. I just needed to clear the interval in the return from the effect.

Test Plan: View runs for a pipeline, navigate to a run. Verify that React warnings don't occur.

Reviewers: bengotow, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D5615