Only send a list of steps down to the backfill graphql mutation when a subset of steps has been selected
Summary:
Somebody in #general reported that they were getting an InvalidStepError when launching a backfill from graphql, but weren't getting that error when kicking off a backfill in the CLI.

Still getting to the bottom of why exactly that would be happening, but while trying to figure it I noticed that we always send the full list of steps down whenever you kick off a backfill, even when you haven't specified a step subset. I think in https://dagster.phacility.com/D4809 that was changed? Generally it should be harmless (and I still want to figure out why this person in #general is running into an InvalidStepError - that should only happen when there's some step in the UI that isn't actually in the execution plan when you go to launch the run), but there's no need to list out all the steps in the common case when you want to do a regular backfill with no step subset,

Test Plan:
Kick off a normal backfill in dagit, watch it complete successfully and the steps get filled in,
Kick off a re-execution backfill with a step subset, watch only those steps get re-executed.

Reviewers: prha, sandyryza, dish, bengotow

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D5695