Move raise_interrupts_directly into user_code_error_boundary
Summary:
Any time we run user code during execution, there's a chance it will hang the execution, but we're currently disabling interrupts during most of execution. Instead, move the logic for turning interrupts back on into user_code_error_boundary.

As part of this, we need to make sure that we don't yield while holding a user_code_error_boundary, which is probably a good idea anyway.

Test Plan: Integration

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D5708