[dagit] Use termination/deletion dialogs on individual run menu
Summary: Use `TerminationDialog` and `DeletionDialog` on the action menu for individual runs on the Runs table.

Test Plan: Kick off a run, view it in Runs page. Verify that "Terminate" and "Delete" are both available, and that they pull up the correct dialogs. Verify that for runs that are already finished, "Terminate" is not rendered as a menu option.

Reviewers: dgibson, prha, johann

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5711