[dagit] Select run rows with Shift+click
Summary: Allow selecting slices of RunTable with shift+click.

Test Plan:
View Run table. Use click and shift+click to check/uncheck rows, verify correct behavior.

Select some rows, then change tabs to a different filter. Verify that the checked runs are carried over if the filter changes but still includes those runs. Verify that if the checked runs are not present in the new filter view, the count (and behavior) of the bulk action menu is still accurate.

Reviewers: bengotow, dgibson, prha, johann

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D5837