[I/o manager cleanup 2/2] make error states consistent between input and output managers
Summary:
In reference to issue https://github.com/dagster-io/dagster/issues/3492.
This diff makes the error states consistent between input and output managers when [1] a resource is provided for a manager key that is not a manager [2] no resource is provided for a manager key.

Test Plan: Added unit tests for failure states

Reviewers: sandyryza, yuhan

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D5849