[dagit] Prefer user timezone on Schedule ticks
Summary:
Show the user's timezone for schedule ticks, with the schedule-specific execution timezone shown in the tooltip. This is the reverse of the current behavior.

Based on discussion today.

I also changed `timestampToString` to use the browser timezone when "Automatic" is set, so that a timezone is at least rendered when included in the format string. This is intended to avoid confusion about execution timezones if the user is using "Automatic" timezone.

Test Plan: View Schedule, verify reversal described above. Verify also that `CST` appears even though I'm on an "Automatic" setting.

Reviewers: prha, dgibson, johann

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5866