Move not_graphql_context_test_suite to root tox file
Summary:
Previously, all the tests that were not part of the dagster-graphql context test matrix were included in the `postgres_tox.ini` file and run as a part of the `dagster-graphql-postgres` BK step. This was due to one test: `test_readonly_variants`, which had a `dagster_postgres` dependency.

This diff updates the tox files to run all tests not part of the test matrix as a part of the `dagster-graphql` step, and adds a `variant` mark that only tests the `test_readonly_variants` in the `dagster-graphql-postgres` step.

Test Plan: bk

Reviewers: dgibson, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D5956