keep AssetStoreOperationType in GRAVEYARD
Summary:
this was my bad from D5863. prior to D5863, we had `ASSET_STORE_OPERATION.op` stored as AssetStoreOperationType in the event log; after that we had it stored as string. The switch caused de/ser issue like https://github.com/dagster-io/dagster/issues/3533
and now as we got rid of ASSET_STORE_OPERATION and we are converting the old logs to the new ones, it'd be safer to capture them both.

This shouldn't affect users much because ASSET_STORE_OPERATION should have only existed if the user opted in to use the asset store before our cherry-pick releases or they have been developing on master.
In https://github.com/dagster-io/dagster/issues/3533 i manually deleted the affected runs. but //if// there is an affected user, it'd be pretty rough experience for both investigating the root cause and fixing it.

Test Plan:
before
{F500696}

after
{F500697}

Reviewers: schrockn, sandyryza, cdecarolis

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D5990